rename component_base file
