+++ stonks/components/base_component.py
