+++ stonks/components/base_fetcher.py
@@ -4,7 +4,7 @@ from contextlib import closing
 import requests
 import click
 
-from stonks.components.component_base import ComponentBase
+from stonks.components.base_component import ComponentBase
 
 # Click progressbar settings for the main request loop
 PROGRESS_SETTINGS = {
